% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{active_hepatitis}
\alias{active_hepatitis}
\title{Chronic active hepatitis}
\format{
A data frame with 44 rows and 3 variables:
\describe{
\item{\code{treatment}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
active_hepatitis
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
