% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bladder}
\alias{bladder}
\title{Recurrence of bladder cancer}
\format{
A data frame with 86 rows and 6 variables:
\describe{
\item{\code{patient}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
\item{\code{treat}}{integer COLUMN_DESCRIPTION}
\item{\code{init}}{integer COLUMN_DESCRIPTION}
\item{\code{size}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
bladder
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
