% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bone_marrow}
\alias{bone_marrow}
\title{Bone marrow transplantation}
\format{
A data frame with 37 rows and 9 variables:
\describe{
\item{\code{patient}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
\item{\code{rage}}{integer COLUMN_DESCRIPTION}
\item{\code{dage}}{integer COLUMN_DESCRIPTION}
\item{\code{type}}{integer COLUMN_DESCRIPTION}
\item{\code{preg}}{integer COLUMN_DESCRIPTION}
\item{\code{index}}{double COLUMN_DESCRIPTION}
\item{\code{gvhd}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
bone_marrow
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
