% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bone_marrow_tx}
\alias{bone_marrow_tx}
\title{Patient outcome following bone marrow transplantation}
\format{
A data frame with 2204 rows and 9 variables:
\describe{
\item{\code{id}}{integer COLUMN_DESCRIPTION}
\item{\code{leukaemia}}{character COLUMN_DESCRIPTION}
\item{\code{age}}{character COLUMN_DESCRIPTION}
\item{\code{match}}{integer COLUMN_DESCRIPTION}
\item{\code{tcell}}{integer COLUMN_DESCRIPTION}
\item{\code{ptime}}{integer COLUMN_DESCRIPTION}
\item{\code{pcens}}{integer COLUMN_DESCRIPTION}
\item{\code{rdtime}}{integer COLUMN_DESCRIPTION}
\item{\code{rdcens}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
bone_marrow_tx
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
