% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast_rfs}
\alias{breast_rfs}
\title{Recurrence free survival in breast cancer patients}
\format{
A data frame with 686 rows and 11 variables:
\describe{
\item{\code{id}}{integer COLUMN_DESCRIPTION}
\item{\code{treat}}{integer COLUMN_DESCRIPTION}
\item{\code{age}}{integer COLUMN_DESCRIPTION}
\item{\code{men}}{integer COLUMN_DESCRIPTION}
\item{\code{size}}{integer COLUMN_DESCRIPTION}
\item{\code{grade}}{integer COLUMN_DESCRIPTION}
\item{\code{nodes}}{integer COLUMN_DESCRIPTION}
\item{\code{prog}}{integer COLUMN_DESCRIPTION}
\item{\code{oest}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
breast_rfs
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
