% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collett-package.R
\docType{package}
\name{collett-package}
\alias{collett}
\alias{collett-package}
\title{collett: Datasets from "Modelling Survival Data in Medical Research" by Collett}
\description{
Datasets for the book entitled "Modelling Survival Data in Medical Research" by Collett (2023) \doi{10.1201/9781003282525}. The datasets provide extensive examples of time-to-event data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mclements/collett}
  \item Report bugs at \url{https://github.com/mclements/collett/issues}
}

}
\author{
\strong{Maintainer}: Mark Clements \email{mark.clements@ki.se} (\href{https://orcid.org/0000-0003-4518-5670}{ORCID})

Other contributors:
\itemize{
  \item Enoch Yi-Tung Chen \email{enoch.yitung.chen@ki.se} (\href{https://orcid.org/0000-0003-2448-708X}{ORCID}) [contributor]
}

}
\keyword{internal}
