% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ducks}
\alias{ducks}
\title{Survival of black ducks}
\format{
A data frame with 50 rows and 6 variables:
\describe{
\item{\code{duck}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
\item{\code{age}}{integer COLUMN_DESCRIPTION}
\item{\code{weight}}{integer COLUMN_DESCRIPTION}
\item{\code{length}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
ducks
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
