% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcancer}
\alias{gcancer}
\title{Survival of patients with gastric cancer}
\format{
A data frame with 90 rows and 4 variables:
\describe{
\item{\code{patient}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
\item{\code{treat}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
gcancer
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
