% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{granulomatous}
\alias{granulomatous}
\title{Chronic granulomatous disease}
\format{
A data frame with 128 rows and 12 variables:
\describe{
\item{\code{patient}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
\item{\code{centre}}{integer COLUMN_DESCRIPTION}
\item{\code{treat}}{integer COLUMN_DESCRIPTION}
\item{\code{age}}{integer COLUMN_DESCRIPTION}
\item{\code{sex}}{integer COLUMN_DESCRIPTION}
\item{\code{height}}{double COLUMN_DESCRIPTION}
\item{\code{weight}}{double COLUMN_DESCRIPTION}
\item{\code{pattern}}{integer COLUMN_DESCRIPTION}
\item{\code{cort}}{integer COLUMN_DESCRIPTION}
\item{\code{anti}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
granulomatous
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
