% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HELP}
\alias{HELP}
\title{Health evaluation and linkage to primary care}
\format{
A data frame with 447 rows and 7 variables:
\describe{
\item{\code{subject}}{integer COLUMN_DESCRIPTION}
\item{\code{days}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
\item{\code{age}}{integer COLUMN_DESCRIPTION}
\item{\code{gender}}{integer COLUMN_DESCRIPTION}
\item{\code{housing}}{integer COLUMN_DESCRIPTION}
\item{\code{linkage}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
HELP
}
\description{
DATASET_DESCRIPTION
}
\details{
Collett (2023) defines this dataset as "help", however
that leads to issues with using R's help system. We have
changed the dataset name to "HELP".
}
\keyword{datasets}
