% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{illustration}
\alias{illustration}
\title{A numerical illustration}
\format{
A data frame with 37 rows and 4 variables:
\describe{
\item{\code{a}}{integer COLUMN_DESCRIPTION}
\item{\code{b}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
illustration
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
