% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kidneytx}
\alias{kidneytx}
\title{Comparisons between kidney transplant centres}
\format{
A data frame with 1439 rows and 9 variables:
\describe{
\item{\code{patient}}{integer COLUMN_DESCRIPTION}
\item{\code{centre}}{integer COLUMN_DESCRIPTION}
\item{\code{tsurv}}{integer COLUMN_DESCRIPTION}
\item{\code{tcens}}{integer COLUMN_DESCRIPTION}
\item{\code{dage}}{integer COLUMN_DESCRIPTION}
\item{\code{dtype}}{integer COLUMN_DESCRIPTION}
\item{\code{rage}}{integer COLUMN_DESCRIPTION}
\item{\code{diab}}{integer COLUMN_DESCRIPTION}
\item{\code{cit}}{double COLUMN_DESCRIPTION}
}
}
\usage{
kidneytx
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
