% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lbrdata0}
\alias{lbrdata0}
\title{Data from a cirrhosis study (lbr data)}
\format{
A data frame with 42 rows and 3 variables:
\describe{
\item{\code{patient}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{lbr}}{double COLUMN_DESCRIPTION}
}
}
\usage{
lbrdata0
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
