% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leukaemia}
\alias{leukaemia}
\title{Bone marrow transplantation in the treatment of leukaemia}
\format{
A data frame with 23 rows and 8 variables:
\describe{
\item{\code{patient}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
\item{\code{group}}{integer COLUMN_DESCRIPTION}
\item{\code{page}}{integer COLUMN_DESCRIPTION}
\item{\code{dage}}{integer COLUMN_DESCRIPTION}
\item{\code{precovery}}{integer COLUMN_DESCRIPTION}
\item{\code{ptime}}{character COLUMN_DESCRIPTION}
}
}
\usage{
leukaemia
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
