% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liver}
\alias{liver}
\title{Survival of liver transplant recipients}
\format{
A data frame with 1761 rows and 7 variables:
\describe{
\item{\code{patient}}{integer COLUMN_DESCRIPTION}
\item{\code{age}}{integer COLUMN_DESCRIPTION}
\item{\code{gender}}{integer COLUMN_DESCRIPTION}
\item{\code{disease}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
\item{\code{cof}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
liver
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
