% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liver_counting}
\alias{liver_counting}
\title{Data from a cirrhosis study (in counting process format)}
\format{
A data frame with 54 rows and 7 variables:
\describe{
\item{\code{patient}}{integer COLUMN_DESCRIPTION}
\item{\code{start}}{integer COLUMN_DESCRIPTION}
\item{\code{stop}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
\item{\code{treat}}{integer COLUMN_DESCRIPTION}
\item{\code{age}}{integer COLUMN_DESCRIPTION}
\item{\code{lbrt}}{double COLUMN_DESCRIPTION}
}
}
\usage{
liver_counting
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
