% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{livertx}
\alias{livertx}
\title{Time to death while waiting for a liver transplant}
\format{
A data frame with 281 rows and 7 variables:
\describe{
\item{\code{patient}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
\item{\code{age}}{integer COLUMN_DESCRIPTION}
\item{\code{gender}}{integer COLUMN_DESCRIPTION}
\item{\code{bmi}}{double COLUMN_DESCRIPTION}
\item{\code{ukeld}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
livertx
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
