% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammary}
\alias{mammary}
\title{Recurrence of mammary tumours in female rats}
\format{
A data frame with 254 rows and 4 variables:
\describe{
\item{\code{rat}}{integer COLUMN_DESCRIPTION}
\item{\code{treatment}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{double COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
mammary
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
