% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melanoma}
\alias{melanoma}
\title{Survival times of patients with melanoma}
\format{
A data frame with 30 rows and 4 variables:
\describe{
\item{\code{age}}{integer COLUMN_DESCRIPTION}
\item{\code{treatment}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
melanoma
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
