% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mice}
\alias{mice}
\title{Survival of laboratory mice}
\format{
A data frame with 181 rows and 3 variables:
\describe{
\item{\code{environment}}{integer COLUMN_DESCRIPTION}
\item{\code{causeofdeath}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
mice
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
