% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myeloma}
\alias{myeloma}
\title{Survival of multiple myeloma patients}
\format{
A data frame with 48 rows and 10 variables:
\describe{
\item{\code{patient}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
\item{\code{age}}{integer COLUMN_DESCRIPTION}
\item{\code{sex}}{integer COLUMN_DESCRIPTION}
\item{\code{bun}}{integer COLUMN_DESCRIPTION}
\item{\code{ca}}{integer COLUMN_DESCRIPTION}
\item{\code{hb}}{double COLUMN_DESCRIPTION}
\item{\code{pcells}}{integer COLUMN_DESCRIPTION}
\item{\code{protein}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
myeloma
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
