% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ovarian}
\alias{ovarian}
\title{Chemotherapy in ovarian cancer patients}
\format{
A data frame with 26 rows and 7 variables:
\describe{
\item{\code{patient}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
\item{\code{treat}}{integer COLUMN_DESCRIPTION}
\item{\code{age}}{integer COLUMN_DESCRIPTION}
\item{\code{rdisease}}{integer COLUMN_DESCRIPTION}
\item{\code{perf}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
ovarian
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
