% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prostatic}
\alias{prostatic}
\title{Comparison of two treatments for prostatic cancer}
\format{
A data frame with 38 rows and 8 variables:
\describe{
\item{\code{patient}}{integer COLUMN_DESCRIPTION}
\item{\code{treatment}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
\item{\code{age}}{integer COLUMN_DESCRIPTION}
\item{\code{shb}}{double COLUMN_DESCRIPTION}
\item{\code{size}}{integer COLUMN_DESCRIPTION}
\item{\code{index}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
prostatic
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
