% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pulmonary}
\alias{pulmonary}
\title{Pulmonary metastasis}
\format{
A data frame with 11 rows and 1 variables:
\describe{
\item{\code{time}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
pulmonary
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
