% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tamoxifen}
\alias{tamoxifen}
\title{Clinical trial of tamoxifen in breast cancer patients}
\format{
A data frame with 641 rows and 18 variables:
\describe{
\item{\code{id}}{integer COLUMN_DESCRIPTION}
\item{\code{treat}}{integer COLUMN_DESCRIPTION}
\item{\code{age}}{integer COLUMN_DESCRIPTION}
\item{\code{size}}{double COLUMN_DESCRIPTION}
\item{\code{hist}}{integer COLUMN_DESCRIPTION}
\item{\code{hr}}{integer COLUMN_DESCRIPTION}
\item{\code{hb}}{integer COLUMN_DESCRIPTION}
\item{\code{andis}}{integer COLUMN_DESCRIPTION}
\item{\code{lsurv}}{integer COLUMN_DESCRIPTION}
\item{\code{ls}}{integer COLUMN_DESCRIPTION}
\item{\code{asurv}}{integer COLUMN_DESCRIPTION}
\item{\code{as}}{integer COLUMN_DESCRIPTION}
\item{\code{dsurv}}{integer COLUMN_DESCRIPTION}
\item{\code{ds}}{integer COLUMN_DESCRIPTION}
\item{\code{msurv}}{integer COLUMN_DESCRIPTION}
\item{\code{ms}}{integer COLUMN_DESCRIPTION}
\item{\code{tsurv}}{integer COLUMN_DESCRIPTION}
\item{\code{ts}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
tamoxifen
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
