% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tplant}
\alias{tplant}
\title{Survival following kidney transplantation}
\format{
A data frame with 434 rows and 7 variables:
\describe{
\item{\code{patient}}{integer COLUMN_DESCRIPTION}
\item{\code{donor}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
\item{\code{age}}{integer COLUMN_DESCRIPTION}
\item{\code{diabetes}}{integer COLUMN_DESCRIPTION}
\item{\code{cit}}{double COLUMN_DESCRIPTION}
}
}
\usage{
tplant
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
