% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ulcer}
\alias{ulcer}
\title{Recurrence of an ulcer}
\format{
A data frame with 43 rows and 6 variables:
\describe{
\item{\code{patient}}{integer COLUMN_DESCRIPTION}
\item{\code{age}}{integer COLUMN_DESCRIPTION}
\item{\code{duration}}{integer COLUMN_DESCRIPTION}
\item{\code{treatment}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{integer COLUMN_DESCRIPTION}
\item{\code{result}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
ulcer
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
