% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valve}
\alias{valve}
\title{Survival following aortic valve replacement}
\format{
A data frame with 988 rows and 11 variables:
\describe{
\item{\code{id}}{integer COLUMN_DESCRIPTION}
\item{\code{futime}}{double COLUMN_DESCRIPTION}
\item{\code{status}}{integer COLUMN_DESCRIPTION}
\item{\code{time}}{double COLUMN_DESCRIPTION}
\item{\code{lvmi}}{double COLUMN_DESCRIPTION}
\item{\code{age}}{integer COLUMN_DESCRIPTION}
\item{\code{sex}}{integer COLUMN_DESCRIPTION}
\item{\code{redo}}{integer COLUMN_DESCRIPTION}
\item{\code{emerg}}{integer COLUMN_DESCRIPTION}
\item{\code{dm}}{integer COLUMN_DESCRIPTION}
\item{\code{type}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
valve
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
