% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{active_hepatitis}
\alias{active_hepatitis}
\title{Chronic active hepatitis}
\format{
A data frame with 44 rows and 3 variables:
\describe{
\item{\code{treatment}}{integer treatment (1=prednisolone, 2=control)}
\item{\code{time}}{integer survival time from admission to study (months)}
\item{\code{status}}{integer event indicator (1=event, 0=right censored)}
}
}
\usage{
active_hepatitis
}
\description{
Clinical trial of 44 patients with chronic active hepatitis randomised to either the drug prednisolone or an untreated control group.
}
\details{
See Collett (2023)
}
\keyword{datasets}
