% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bladder}
\alias{bladder}
\title{Recurrence of bladder cancer}
\format{
A data frame with 86 rows and 6 variables:
\describe{
\item{\code{patient}}{integer patient number (1-86)}
\item{\code{time}}{integer survival time in months}
\item{\code{status}}{integer status of patient (0=censored, 1=recurrence)}
\item{\code{treat}}{integer treatment group (1=placebo, 2=thiotepa)}
\item{\code{init}}{integer initial number of tumours}
\item{\code{size}}{integer diameter of larger initial tumour in cm}
}
}
\usage{
bladder
}
\description{
Placebo controlled trial of bladder cancer patients randomised to thiopeta or to placebo
}
\details{
See Collett (2023)
}
\keyword{datasets}
