% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bone_marrow_tx}
\alias{bone_marrow_tx}
\title{Patient outcome following bone marrow transplantation}
\format{
A data frame with 2204 rows and 9 variables:
\describe{
\item{\code{id}}{integer patient id}
\item{\code{leukaemia}}{character type of leukaemia (CML,ALL,AML)}
\item{\code{age}}{character age group of patient in years (<=20, 21-40, >40))}
\item{\code{match}}{integer indicator for whether there was a donor gender match (0=no, 1=yes)}
\item{\code{tcell}}{integer indicator for whether there was T-cell depletion (1=yes, n=no)}
\item{\code{ptime}}{integer time to platelet recovery (days)}
\item{\code{pcens}}{integer event indicator for platelet recovery (1=event, 0=censored)}
\item{\code{rdtime}}{integer time to relapse of death (days)}
\item{\code{rdcens}}{integer event indicator for relapse or death (1=event, 0=censored)}
}
}
\usage{
bone_marrow_tx
}
\description{
Patient outcome following bone marrow transplantation
}
\details{
See Collett (2023)
}
\keyword{datasets}
