% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collett-package.R
\docType{package}
\name{collett-package}
\alias{collett}
\alias{collett-package}
\title{collett: Datasets and Examples from "Modelling Survival Data in Medical Research" by Collett}
\description{
Datasets and Examples from "Modelling Survival Data in Medical Research" by Collett.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mclements/collett}
  \item Report bugs at \url{https://github.com/mclements/collett/issues}
}

}
\author{
\strong{Maintainer}: Mark Clements \email{mark.clements@ki.se} (\href{https://orcid.org/0000-0003-4518-5670}{ORCID})

}
\keyword{internal}
