% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dialysis}
\alias{dialysis}
\title{Infection in patients on dialysis}
\format{
A data frame with 13 rows and 5 variables:
\describe{
\item{\code{patient}}{integer patient id}
\item{\code{time}}{integer time to infection (days)}
\item{\code{status}}{integer event indicator (0=censored, 1=infection)}
\item{\code{age}}{integer age in years}
\item{\code{sex}}{integer sex of the patient (1=male, 2=female)}
}
}
\usage{
dialysis
}
\description{
Time from dialysis to infection for patients with diseases of the kidney.
}
\details{
See Collett (2023)
}
\keyword{datasets}
