% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ducks}
\alias{ducks}
\title{Survival of black ducks}
\format{
A data frame with 50 rows and 6 variables:
\describe{
\item{\code{duck}}{integer duck indicator}
\item{\code{time}}{integer survival time in days}
\item{\code{status}}{integer status of bird (0=alive or missing, 1=dead)}
\item{\code{age}}{integer age group (0=hatch-year bird, 1=bird aged >= 1 year)}
\item{\code{weight}}{integer weight of bird in g}
\item{\code{length}}{integer length of wing in mm}
}
}
\usage{
ducks
}
\description{
Black ducks, Anas rubripes, were followed the US Fish and Wildlife Service.
}
\details{
See Collett (2023)
}
\keyword{datasets}
