% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcancer}
\alias{gcancer}
\title{Survival of patients with gastric cancer}
\format{
A data frame with 90 rows and 4 variables:
\describe{
\item{\code{patient}}{integer patient id}
\item{\code{time}}{integer survival time in days}
\item{\code{status}}{integer event indicator (0=censored, 1=dead)}
\item{\code{treat}}{integer treatment arm (0=chemotherapy alone, 1=chemotherapy and radiotherapy)}
}
}
\usage{
gcancer
}
\description{
Survival of patients with gastric cancer
}
\details{
See Collett (2023)
}
\keyword{datasets}
