% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{granulomatous}
\alias{granulomatous}
\title{Chronic granulomatous disease}
\format{
A data frame with 128 rows and 12 variables:
\describe{
\item{\code{patient}}{integer patient number (1-128)}
\item{\code{time}}{integer time to first infection (days)}
\item{\code{status}}{integer status of patient (0=censored, 1=infection)}
\item{\code{centre}}{integer treatment centre; see Collett (2023, page 504)}
\item{\code{treat}}{integer treatment group (0=placebo, 1=interferon)}
\item{\code{age}}{integer age in years}
\item{\code{sex}}{integer sex (1=male, 2=female)}
\item{\code{height}}{double height in cm}
\item{\code{weight}}{double weight in kg}
\item{\code{pattern}}{integer pattern of inheritance (1=X-linked, 2=autosomal recessive)}
\item{\code{cort}}{integer use of corticosteroids at trial entry (1=used, 2=not used)}
\item{\code{anti}}{integer Use of antibiotics at trial entry (1=used, 2=not used)}
}
}
\usage{
granulomatous
}
\description{
Trial comparing interferon with a placebo.
}
\details{
See Collett (2023)
}
\keyword{datasets}
