% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{illustration}
\alias{illustration}
\title{A numerical illustration}
\format{
A data frame with 37 rows and 4 variables:
\describe{
\item{\code{a}}{integer factor a}
\item{\code{b}}{integer factor b}
\item{\code{time}}{integer event time}
\item{\code{status}}{integer event status (1=event, 0=right censored)}
}
}
\usage{
illustration
}
\description{
Artificial data on patient survival classified according to factors a and b
}
\details{
See Collett (2023).
}
\keyword{datasets}
