% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lbrdata0}
\alias{lbrdata0}
\title{Data from a cirrhosis study (lbr data)}
\format{
A data frame with 42 rows and 3 variables:
\describe{
\item{\code{patient}}{integer patient id}
\item{\code{time}}{integer date of measurement (days)}
\item{\code{lbr}}{double log bilirubin level}
}
}
\usage{
lbrdata0
}
\description{
DATASET_DESCRIPTION
}
\details{
See Collett (2023)
}
\keyword{datasets}
