% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liver_counting}
\alias{liver_counting}
\title{Data from a cirrhosis study (in counting process format)}
\format{
A data frame with 54 rows and 7 variables:
\describe{
\item{\code{patient}}{integer patient id}
\item{\code{start}}{integer start time (days)}
\item{\code{stop}}{integer stop time (days)}
\item{\code{status}}{integer event indicator (0=censored, 1=uncensored)}
\item{\code{treat}}{integer treatment group (0=placebo, 1=Liverol)}
\item{\code{age}}{integer age of the patient at start of study (years)}
\item{\code{lbrt}}{double logarithm of bilirubin level}
}
}
\usage{
liver_counting
}
\description{
Artificial data
}
\details{
See Collett (2023). Note that the variable for log of bilirubin differs to that for "liverbase".
}
\keyword{datasets}
