% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liverbase}
\alias{liverbase}
\title{Data from a cirrhosis study (baseline)}
\format{
A data frame with 12 rows and 6 variables:
\describe{
\item{\code{patient}}{integer patient id}
\item{\code{time}}{integer survival time in days}
\item{\code{status}}{integer event indicator (0=censored, 1=uncensored)}
\item{\code{age}}{integer age of the patient (years)}
\item{\code{treat}}{integer treatment group (0=placebo, 1=Liverol)}
\item{\code{lbr}}{double logarithm of bilirubin level}
}
}
\usage{
liverbase
}
\description{
Articial data
}
\details{
See Collett (2023)
}
\keyword{datasets}
