% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lung}
\alias{lung}
\title{Survival of patients registered for a lung transplant}
\format{
A data frame with 196 rows and 7 variables:
\describe{
\item{\code{patient}}{integer patient id}
\item{\code{time}}{integer time from registration to the earlist of removal from list, last known follow-up date, 30 April 2012, or death (days)}
\item{\code{status}}{integer event indicator (0=censored, 1=dead)}
\item{\code{age}}{integer age in years}
\item{\code{gender}}{integer gender (1=male, 2=female)}
\item{\code{bmi}}{double body mass index}
\item{\code{disease}}{integer disease (1=COPD, 2=fibrosis, 3=suppurative, 4=other)}
}
}
\usage{
lung
}
\description{
Survival of patients registered for a lung transplant
}
\details{
See Collett (2023)
}
\keyword{datasets}
