% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammary}
\alias{mammary}
\title{Recurrence of mammary tumours in female rats}
\format{
A data frame with 254 rows and 4 variables:
\describe{
\item{\code{rat}}{integer id for each rat}
\item{\code{treatment}}{integer treatment arm indicator (1=treatment, 0=control)}
\item{\code{time}}{double follow-up time (days)}
\item{\code{status}}{integer recurrence indicator (0=no, 1=yes)}
}
}
\usage{
mammary
}
\description{
This is an animal experiment to compare the use of retinyl acetate (related to vitamin A) across the study (treatment) to treatment with retinyl acetate to 60 days and then no further treatment (control). The female rats all had mammary tumours.
}
\details{
See Collett (2023)
}
\keyword{datasets}
