% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melanoma}
\alias{melanoma}
\title{Survival times of patients with melanoma}
\format{
A data frame with 30 rows and 4 variables:
\describe{
\item{\code{age}}{integer age group (1=21-44, 2=41-60, 3=61+)}
\item{\code{treatment}}{integer treatment arm (1=BCG, 2=C. parvum)}
\item{\code{time}}{integer survival time (months)}
\item{\code{status}}{integer event indicator (0=censored, 1=dead)}
}
}
\usage{
melanoma
}
\description{
Comparing two immunotherapy treatments for patients with melanoma
}
\details{
See Collett (2023)
}
\keyword{datasets}
