% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ovarian}
\alias{ovarian}
\title{Chemotherapy in ovarian cancer patients}
\format{
A data frame with 26 rows and 7 variables:
\describe{
\item{\code{patient}}{integer identifer}
\item{\code{time}}{integer survival time from randomisation in days}
\item{\code{status}}{integer event indicator (0=right censored, 1=event)}
\item{\code{treat}}{integer treatment (1=single, 2=combined)}
\item{\code{age}}{integer age of patients in years}
\item{\code{rdisease}}{integer extent of residual disease (1=incomplete, 2=complete)}
\item{\code{perf}}{integer performance status (1=good, 2=poor)}
}
}
\usage{
ovarian
}
\description{
Trial for treatment of ovarian cancer patients comparing cyclophosphamide alone with cyclophosphamide combined with adriamycin.
}
\details{
See Collett (2023)
}
\keyword{datasets}
