% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pulmonary}
\alias{pulmonary}
\title{Pulmonary metastasis}
\format{
A data frame with 11 rows and 1 variables:
\describe{
\item{\code{time}}{integer survival time from pulmonary metastasis to death in months}
}
}
\usage{
pulmonary
}
\description{
A very simple dataset with no censoring
}
\details{
See Collett (2023)
}
\keyword{datasets}
