% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valve}
\alias{valve}
\title{Survival following aortic valve replacement}
\format{
A data frame with 988 rows and 11 variables:
\describe{
\item{\code{id}}{integer patient id}
\item{\code{futime}}{double total follow-up time from date of surgery (years)}
\item{\code{status}}{integer event indicator (0=censored, 1=death)}
\item{\code{time}}{double time of LVMI measurement after surgery (years)}
\item{\code{lvmi}}{double standardised LVMI}
\item{\code{age}}{integer age of patient in years}
\item{\code{sex}}{integer sex of patient (0=male, 1=female)}
\item{\code{redo}}{integer previous cardiac surgery (0=no, 1=yes)}
\item{\code{emerg}}{integer operative urgency (0=elective, 1=urgent or emergency)}
\item{\code{dm}}{integer preoperative diabetes mellitus (0=no, 1=yes)}
\item{\code{type}}{integer type of valve (1=human tissue, 2=porcine tissue)}
}
}
\usage{
valve
}
\description{
Patients following an aortic valve replacement are measured for left ventricular mass index (LVMI).
}
\details{
See Collett (2023)
}
\keyword{datasets}
